Move query streaming logic to its own file
After the refactor in the previous commit, this is somewhat independent of the query routes.
